+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # CHANGELOG
 
-## 6.11.2016
-    - Release of 1.0.0 version
+## next release date: 2.0.0 release
+    - [new] Implements configuration with using DI cotainer
+    - [brake] Method `getTransport` now is NOT protected
+    - [new] Method `getTransport` now is public
+    - [brake] Method `getPublisher` now is NOT protected
+    - [new] Method `getPublisher` now is public
+    - [new] `MessageValidator` for `Publisher` now is configurable
+    - [new] You can specify DI container in config 
+    
+## 6.11.2016: 1.0.0 release